Linux/ESPNOW: remove delivery confirmation
Removes delivery confirmation (processing of 802.11 ACKs), because
waiting for them with timeout blocks the transmission for too long
and isn't reliable.
Also removes retransmissions, as they are being automatically
handled by driver.

As Linux port of ESP-NOW is mainly focused on bridge nodes and they
don't implement any retransmission logic anyway, it's not a big
problem for now.